Fix: Adjust I and L after razor assignment # (#517)
* refactor: Improve clarity of code, add comments

Added comments, changed some misleading variable names, slightly modified how the
updated protein start and end positions dervied from the database are stored and reused.

* fix: I and L in peptide occasionally mixed up after promoting proteins

When importing pepXML files in philosopher, a different protein can be selected from the
alternative proteins as the reported protein (swiss prot entries preferred, decoys are
removed). When this happens, the previous and next amino acids, as well as the protein
start and end positions of a peptide are corrected. However, the peptide sequence is
not updated. In rare cases this can lead to the reporting of wrong peptide sequences
because I/L are treated as indistuingishable for peptide spectrum matching, but there
can be differences of I and L in an otherwise identical peptide sequence derived from
two different proteins.

This commit introduces changes to `lib/rep/updater/UpdateLayerswithDatabase` function,
which now also updates peptide and modified peptide sequences with the observed ones
from a protein database (FASTA).